adding a new test and some new fields to blogs

+++ src/repositories/blogs-repo.ts
@@ -1,18 +1,23 @@
 // import { v4 as uuidv4 } from "uuid";
 import { client } from "./db";
 import { ObjectId } from "mongodb";
+import { blogs } from "./db";
 
 export type Blog = {
   name: string;
   description: string;
   websiteUrl: string;
+  //TODO this won't be optional in the future
+  isMembership?: boolean;
 };
-
-const blogs = client.db("samurai").collection("blogs");
 
 export const blogsRepository = {
   async create(blog: Blog) {
-    const result = await blogs.insertOne(blog);
+    const result = await blogs.insertOne({
+      ...blog,
+      isMembership: false,
+      createdAt: new Date(),
+    });
     return result.insertedId.toString();
   },
   async findAll() {
@@ -20,15 +25,27 @@ export const blogsRepository = {
   },
 
   async findById(id: string) {
+    if (!ObjectId.isValid(id)) {
+      return null;
+    }
     return blogs.findOne({ _id: new ObjectId(id) });
   },
   async deleteById(id: string) {
+    if (!ObjectId.isValid(id)) {
+      return false;
+    }
     const result = await blogs.deleteOne({ _id: new ObjectId(id) });
     return result.deletedCount === 1;
   },
   async updateById(id: string, blog: Blog) {
-    const result = await blogs.replaceOne({ _id: new ObjectId(id) }, blog);
-    return result.modifiedCount === 1;
+    if (!ObjectId.isValid(id)) {
+      return false;
+    }
+    const result = await blogs.updateOne(
+      { _id: new ObjectId(id) },
+      { $set: blog }
+    );
+    return result.matchedCount === 1;
   },
   async deleteAll() {
     //drop the blogs collection
